:feature: circlular minimap
