+++ include/cub.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   cub.h                                              :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: aboudoun <aboudoun@student.42.fr>          +#+  +:+       +#+        */
+/*   By: majjig <majjig@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/09/30 16:22:17 by aboudoun          #+#    #+#             */
-/*   Updated: 2022/10/05 14:56:07 by aboudoun         ###   ########.fr       */
+/*   Updated: 2022/10/06 14:02:11 by majjig           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -150,5 +150,6 @@ double			get_vertical(t_data *data, double angle, int count);
 int				get_rgb(char *num);
 void			init_data(t_data *data);
 void			init_images(t_data *data);
+void			circlular_minimap(int *img);
 
 #endif
